+++ libraries/typescript/packages/inspector/package.json
@@ -24,6 +24,10 @@
     ".": {
       "types": "./dist/server/index.d.ts",
       "import": "./dist/server/index.js"
+    },
+    "./client": {
+      "types": "./dist/client/index.d.ts",
+      "import": "./dist/client/index.js"
     }
   },
   "main": "./dist/server/index.js",
@@ -39,8 +43,9 @@
     "dev:client": "rm -rf node_modules/.vite && vite --port 3000",
     "dev:server": "VITE_DEV=true tsx watch src/server/server.ts",
     "dev:standalone": "tsx watch src/server/server.ts",
-    "build": "npm run build:client && npm run build:server && npm run build:cli",
+    "build": "npm run build:client && npm run build:client-exports && npm run build:server && npm run build:cli",
     "build:client": "vite build",
+    "build:client-exports": "tsup src/client/index.ts --format esm --out-dir dist/client --tsconfig tsconfig.client.json --external react --external react-dom --external lucide-react --no-splitting && tsc -p tsconfig.client.json --emitDeclarationOnly --declaration",
     "build:server": "tsup src/server/*.ts --format esm --out-dir dist/server && tsc -p tsconfig.server.json --emitDeclarationOnly --declaration",
     "build:cli": "tsup src/server/cli.ts --format esm --out-dir dist/",
     "prepublishOnly": "pnpm --filter mcp-use build && pnpm build",
