+++ libraries/typescript/packages/inspector/src/client/components/ui/not-found.tsx
@@ -1,4 +1,3 @@
-import type React from "react";
 import { cn } from "@/client/lib/utils";
 
 interface NotFoundProps {
