+++ libraries/typescript/packages/inspector/src/client/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Client-side exports for inspector package
+ */
+
+export { AddToClientDropdown } from "./components/AddToClientDropdown.js";
+export * from "./utils/mcpClientUtils.js";
