+++ libraries/typescript/packages/inspector/src/client/utils/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Utility functions for MCP client configurations
+ */
+
+export * from "./mcpClientUtils.js";
