+++ libraries/typescript/packages/mcp-use/src/client/browser.ts
@@ -68,6 +68,7 @@ export class BrowserMCPClient extends BaseMCPClient {
       elicitationCallback,
       disableSseFallback,
       preferSse,
+      clientInfo,
     } = serverConfig;
 
     if (!url) {
@@ -85,6 +86,7 @@ export class BrowserMCPClient extends BaseMCPClient {
       elicitationCallback, // ← Pass elicitation callback to connector
       disableSseFallback, // ← Disable automatic SSE fallback
       preferSse, // ← Use SSE transport directly
+      clientInfo, // ← Pass client info (name, version) to connector
     };
 
     // Debug: Log if clientOptions are being passed
