+++ libraries/typescript/packages/mcp-use/src/utils/index.ts
@@ -4,5 +4,4 @@
  */
 
 export * from "./favicon-detector.js";
-export * from "./mcpClientUtils.js";
 export * from "./proxy-config.js";
